requirements.txt
annotated-types==0.7.0
anyio==4.9.0
babel==2.17.0
black==24.8.0
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
click==8.1.8
colorama==0.4.6
coverage==7.10.1
cryptography==45.0.5
Deprecated==1.2.18
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
environs==14.1.1
fastapi==0.115.12
fastapi-cli==0.0.7
fastapi-limiter==0.1.6
filelock==3.18.0
flake8==7.1.1
Flake8-pyproject==1.2.3
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.10
idna==3.10
iniconfig==2.1.0
isort==6.0.0b2
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.5.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.6.0
pre-commit==3.8.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.2.0
rich==14.0.0
rich-toolkit==0.14.5
shellingham==1.5.4
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.46.2
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.31.1
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.3
